#13 initialization of domain completed

+++ backend/src/test/java/com/desofs/backend/domain/valueobjects/MoneyAmountTest.java
@@ -3,6 +3,8 @@ package com.desofs.backend.domain.valueobjects;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class MoneyAmountTest {
@@ -10,16 +12,16 @@ class MoneyAmountTest {
     @Test
     @DisplayName("Test create method with valid money amount")
     void testCreateValidMoneyAmount() {
-        MoneyAmount moneyAmount = MoneyAmount.create(100.50f);
+        MoneyAmount moneyAmount = MoneyAmount.create(BigDecimal.valueOf(100.50));
         assertNotNull(moneyAmount);
-        assertEquals(100.50f, moneyAmount.getValue());
+        assertEquals(BigDecimal.valueOf(100.50), moneyAmount.getValue());
     }
 
     @Test
     @DisplayName("Test create method with negative money amount")
     void testCreateNegativeMoneyAmount() {
         IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> MoneyAmount.create(-100.50f));
+                () -> MoneyAmount.create(BigDecimal.valueOf(-100.50)));
         assertEquals("MoneyAmount must be positive.", exception.getMessage());
     }
 }
